requirements.txt
pip
setuptools
django
django-imagekit
Pillow-PIL